arreglo img delete

+++ src/models/adminModel.php
@@ -35,7 +35,7 @@ function deleteAdmin(mysqli $conexion, string $bbdd, int $id)
    $row = mysqli_fetch_assoc($result);
    $foto = $row['foto'] ?? null;
 
-   if ($foto) {
+   if ($foto && $foto != "no_photo.png") {
       $filePath = __DIR__ . "/../../media/img/" . $foto;
       if (file_exists($filePath)) {
          unlink($filePath);
