+++ src/models/estudianteModel.php
@@ -56,7 +56,7 @@ function deleteStudent(mysqli $conexion, string $bbdd, int $id)
    $row = mysqli_fetch_assoc($result);
    $foto = $row['foto'] ?? null;
 
-   if ($foto) {
+   if ($foto && $foto != "no_photo.png") {
       $filePath = __DIR__ . "/../../media/img/" . $foto;
       if (file_exists($filePath)) {
          unlink($filePath);
